chore: refactor vehicles controller
