+++ app/views/api/v1/hangar_groups/public_index.jbuilder
