+++ app/views/api/v1/hangar_stats/index.jbuilder
