+++ app/views/api/v1/hangar_stats/public_index.jbuilder
