--- app/views/api/v1/vehicles/fleetchart.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @vehicles, partial: 'api/v1/vehicles/minimal', as: :vehicle
