--- app/views/api/v1/vehicles/public_fleetchart.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @vehicles, partial: 'api/v1/vehicles/minimal_public', as: :vehicle
