+++ app/views/api/v1/vehicles/public_index.jbuilder
